Add files via upload
GEGL Effects July 22 2023 Release. Bevel got an update that gives it a new sharp bevel mode that is also in GEGL Effects.  

Also, like many other filters GEGL Effects got a update to show up in Gimp 2.99/3's main menu via "gimp:menu-path", and gimp:menu-label. This will only be applied in 2.99/3 and not Gimp 2.10.  2.10 users will have the sharp bevel of course but not the filter appearing in the menu.